Added reference to Yeoman, corrected conflicting "open/create"
Added instructions for installing the skeleton project via Yeoman rather than downloading and unzipping it. The Yeoman generator "generator-aurelia" just grabs the newest version of the skeleton project from GitHub, so it is more or less equivalent.

Also, I recently made a mistake changing the instructions to "open" the index.html instead of "creating" it, since I missed that the user had earlier been instructed to delete the file ahead of time. I've corrected it back to "create", moved the initial deletion instructions into a new paragraph, and reworded them a bit to accommodate people who would rather examine the files first.