Update 5. cheat-sheet.md
The solution proposed and documented here https://github.com/aurelia/templating/issues/261 did not work for me. This was due to webpack not supporting dynamic compose https://github.com/aurelia/skeleton-navigation/issues/810. @jods4 proposed to bind to a property on the viewmodel that uses `PLATFORM.moduleName` which does then allow the solution to work. This update is to just give a heads up to anyone reading the cheatsheet that is using webpack.