Update dispatchify example to be stand-alone
The example relies too much on re-using files from previous parts of the page.

For example: setting up the app.html, having configured the store, etc.

The problem is that the non of the previous files setup the action
`addFramework`, yet the definition of this action is in this example.

As the example is meant to be `Store`less the registration of actions should
also be done elsewhere.

Also, while attempting to follow along with this example and refactoring down
the framework list out of `app.html` and into `framework-list` the problem with
inlineViews and template strings containing $ gets caused. See
https://github.com/aurelia/store/issues/57. As the example uses inlineViews I've
also escaped the string template to work correctly.